requirements.txt
# Core RAG
openai>=1.12.0
groq>=0.4.0
langchain>=0.1.6
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Vector Database
qdrant-client>=1.7.1

# Embeddings
sentence-transformers>=2.2.2
numpy>=1.26.0

# UI
streamlit>=1.31.0

# API (for frontend integration)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Optional (for future)
redis>=5.0.1
prometheus-client>=0.19.0
